feat: Add a random key generator
The essence of a random key generator is for having secure keys for sensitive
information that needs to be secured. Security begins with the implementer of
said security and what is security to start with, when a key is weak.

This was why this solution is proposed and it can be used through out the
lifecycle of the application